requirements.txt
tzdata==2025.1
discord.py==2.5.0
requests==2.32.3